02_bootstrap-scripts-python/requirements.txt
ruamel.yaml==0.17.10
jsonpath-ng==1.5.2
deepmerge==0.3.0
